requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
mysql-connector-python
python-dotenv
PyJWT